fliprLicensingApi/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.135
botocore==1.29.135
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
Django==4.2.1
django-cors-headers==3.14.0
django-prometheus==2.3.1
django-redis==5.2.0
django-ses==3.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
kombu==5.2.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.5.5
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
